+++ next/src/components/common/Footer/DesktopFooter.tsx
@@ -1,7 +1,6 @@
 import { Typography } from '@bratislava/component-library'
 
 import { EuFlagSvg } from '@/src/assets/images'
-import Brand from '@/src/components/common/Brand/Brand'
 import HorizontalDivider from '@/src/components/common/Divider/HorizontalDivider'
 import {
   FooterAccessibilityLink,
@@ -11,6 +10,7 @@ import {
   FooterLanguageSwitcher,
   FooterSocialLinks,
 } from '@/src/components/common/Footer/FooterShared'
+import Brand from '@/src/components/common/Logos/Brand'
 import SectionContainer from '@/src/components/layouts/SectionContainer'
 import { useGeneralContext } from '@/src/components/providers/GeneralContextProvider'
 import { isDefined } from '@/src/utils/isDefined'
